+++ SimpleFarm/SimpleFarm.cs
@@ -1,4 +1,7 @@
 ﻿global using HarmonyLib;
+using Shared;
+using System.Reflection;
+using System.Reflection.Emit;
 
 namespace SimpleFarmNS
 {
@@ -6,6 +9,7 @@ namespace SimpleFarmNS
     {
         public void Awake()
         {
+            PatchSafe(typeof(Patch_Greenhouse));
             Logger.Log($"Awake!");
         }
 
@@ -27,23 +31,30 @@ namespace SimpleFarmNS
                         && subprint.ExtraResultCards != null
                         && subprint.ExtraResultCards.Length > 0)
                     {
+                        // replace food trees and bushes with just the food
                         var card_in = subprint.RequiredCards[0];
                         if (subprint.ExtraResultCards.Length == 1)
                         {
-                            if (card_in is not ("stick"))
-                                subprint.ExtraResultCards[0] = card_in;
-                        }
-                        else
-                        {
-                            Array.Resize(ref subprint.ExtraResultCards, subprint.ExtraResultCards.Length - 1);
+                            if (card_in is "stick")
+                                subprint.ExtraResultCards = ["stick", "tree"];
+                            else
+                                subprint.ExtraResultCards = [card_in, card_in];
                         }
 
-                        // if CardsToRemove is null or empty, then the game will destroy all RequiredCard that are not structures
-                        // thus we use a non existing id, which is catched by the game's code
-                        subprint.CardsToRemove = ["removenocard"];
+                        // greenhouse boosts
+                        if (subprint.RequiredCards[1] is "greenhouse")
+                        {
+                            int foodvalue = WorldManager.instance.GetCardPrefab(card_in, false) is Food food ? food.FoodValue : 0;
+                            if (foodvalue == 1)
+                            {
+                                Array.Resize(ref subprint.ExtraResultCards, subprint.ExtraResultCards.Length + 1);
+                                subprint.ExtraResultCards[^1] = card_in;
+                            }
+                            //subprint.Time = Math.Max(10f, subprint.Time - 10f);
+                        }
                     }
 
-                    Logger.Log($"{subprint.RequiredCards?.Join()} => {subprint.ExtraResultCards?.Join()} || {subprint.CardsToRemove?.Join()}");
+                    Logger.Log($"{subprint.RequiredCards?.Join()} => {subprint.ExtraResultCards?.Join()}");
                 }
             }
 
@@ -54,12 +65,34 @@ namespace SimpleFarmNS
         {
             try
             {
+                Logger.Log($"Patching {patch.Name}");
                 Harmony.CreateClassProcessor(patch).Patch();
                 return true;
             } catch (Exception e)
             {
                 Logger.LogException(e.ToString());
                 return false;
+            }
+        }
+    }
+
+    /// <summary>
+    /// Makes garden and farm behave like greenhouses. This is preferable to changing the Subprint, because of spoilage.
+    /// </summary>
+    [HarmonyPatch(typeof(BlueprintGrowth), nameof(BlueprintGrowth.BlueprintComplete))]
+    public static class Patch_Greenhouse
+    {
+        public static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions, ILGenerator generator, MethodBase original)
+        {
+            var data = new TranspilerTool(instructions, generator, original);
+            data.Seek(OpCodes.Ldstr, "greenhouse");
+            data++;
+            data.InsertAfter(patch);
+            return data.Code;
+
+            static bool patch(bool stack, GameCard rootCard)
+            {
+                return stack || rootCard.CardData.Id is "garden" or "farm";
             }
         }
     }