+++ SimpleFarm/changelog.md
@@ -1,4 +1,10 @@
 # Changelog
 
+## [1.0.1]
+- fixed cards getting stuck in greenhouse
+- fixed spoilage on island board
+- changed greenhouse to produce 3 cards, if food value is 1
+- changed stick to produce stick and tree
+
 ## [1.0.0]
 - first release